requirements.txt
streamlit
tensorflow==2.19.0
keras
numpy
pandas
matplotlib
Pillow
scikit-learn
